[hitomi]: hitomi-tools first-run pass
